scripts/requirements.txt
openai>=1.30.0,<2
numpy>=1.26

# For calendario de vencimientos tooling
pandas>=2.2
openpyxl>=3.1
XlsxWriter>=3.2
python-dateutil>=2.9
ics>=0.7

# Google Calendar API
google-api-python-client>=2.131.0
google-auth>=2.33.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1



